requirements.txt
apache-airflow==1.10.9
appdirs==1.4.3
asn1crypto==1.3.0
attrs==19.3.0
awscli==1.17.2
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.11.2
botocore==1.14.2
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
idna==2.8
ijson==2.6.1
isodate==0.6.0
Jinja2<2.11.0,>=2.10.1
jmespath==0.9.4
kubernetes==11.0.0
lxml==4.5.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
networkx==2.4
oscrypto==1.2.0
protobuf==3.6.1
py4j==0.10.7
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.9.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
snowflake-connector-python==2.2.2
snowflake-sqlalchemy==1.2.0
SQLAlchemy==1.3.13
sqlalchemy-redshift==0.7.7
stripe==2.42.0
tabulate==0.8.6
tzlocal<2.0.0,>=1.4
urllib3==1.25.8
zeep==3.4.0
